implemented security

+++ userinfo/src/main/java/com/food/userinfo/service/UserService.java
@@ -6,5 +6,5 @@ import org.springframework.http.ResponseEntity;
 public interface UserService {
     UserDTO addUser(UserDTO userDTO);
 
-    ResponseEntity<UserDTO> fetchUserDetailsById(Integer userId);
+    ResponseEntity<UserDTO> fetchUserDetailsById(Long userId);
 }
